Add an acomm.SyncRequest convenience method
Expose the external port in test coordinator
Use ip instead of hostname for node heartbeat id
Add a ListNodes to clusterconf
Add datatrade provider with DatasetImport task handler
datatrade provider cmd
Linting fixes
Regenerate Readme
Make Quota and Redundancy uint64

+++ providers/clusterconf/bundle_test.go
@@ -20,7 +20,7 @@ func (s *clusterConf) TestGetBundle() {
 	tests := []struct {
 		desc            string
 		id              uint64
-		quota           int
+		quota           uint64
 		dataset         string
 		combinedOverlay bool
 		err             string
